+++ packages/layers/CHANGELOG.md
@@ -1,5 +1,25 @@
 # @vivjs/layers
 
+## 0.20.1
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @vivjs/constants@0.20.1
+  - @vivjs/extensions@0.20.1
+  - @vivjs/loaders@0.20.1
+  - @vivjs/types@0.20.1
+
+## 0.20.0
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @vivjs/constants@0.20.0
+  - @vivjs/extensions@0.20.0
+  - @vivjs/loaders@0.20.0
+  - @vivjs/types@0.20.0
+
 ## 0.19.0
 
 ### Patch Changes